+++ Code/Renderer/textureCreation.cpp
@@ -34,7 +34,7 @@ VkImageView TextureUtilities::createImageView(VkDevice device, VkImage image, Vk
 }
 
 
-void TextureUtilities::createImage(RendererContext rendererHandles, uint32_t width, uint32_t height, VkFormat format,
+void TextureUtilities::createImage(RendererContext rendererHandles, uint64_t width, uint64_t height, VkFormat format,
                                    VkImageTiling tiling,
                                    VkImageUsageFlags usage, VkMemoryPropertyFlags properties, VkImage& image,
                                    VmaAllocation& allocation, uint32_t miplevels, uint32_t araryLayers, bool cubeCompatible)
@@ -44,8 +44,8 @@ void TextureUtilities::createImage(RendererContext rendererHandles, uint32_t wid
     VkImageCreateInfo imageInfo{};
     imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
     imageInfo.imageType = VK_IMAGE_TYPE_2D ;
-    imageInfo.extent.width = width;
-    imageInfo.extent.height = height;
+    imageInfo.extent.width = (uint32_t)width;
+    imageInfo.extent.height = (uint32_t)height;
     imageInfo.extent.depth = 1;
     imageInfo.mipLevels = miplevels;
     imageInfo.arrayLayers = araryLayers;
